product_tracker/requirements.txt
requests==2.22.0
beautifulsoup4==4.10.0
python-dotenv==0.19.2
utils==1.0.1
